refactor: 현재 활성화된 nav 요소를 계산하는 로직을 훅으로 분리
- 여러 페이지에서 재사용하기 위해 해당 로직을 훅으로 분리
